src/pyproject.toml
[tool.poetry]
name = "rss_aggregator_fastapi"
version = "0.1.0"
description = ""
authors = ["sqvall <masster1987@inbox.ru>"]
license = ""

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.14.0"
gunicorn = "^20.1.0"
alembic = "^1.6.5"
asyncpg = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.15.1"
httpx = "^0.18.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
